% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{buildBreakdowns}
\alias{buildBreakdowns}
\title{buildBreakdowns}
\usage{
buildBreakdowns(breakdowns, f)
}
\arguments{
\item{breakdowns}{Breakdowns}

\item{f}{testParam}
}
\description{
buildBreakdowns
}
