% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkTarget.R
\name{checkTarget}
\alias{checkTarget}
\title{Check object target}
\usage{
checkTarget(id, token, n = 100)
}
\arguments{
\item{id}{Either a adset ID or an ad ID.}

\item{token}{A valid token as returned by \code{\link{fbAuthenticate}} or a short-term token from \href{https://developers.facebook.com/tools/explorer}{facebook Graph API Explorer}.}

\item{n}{Number of results to retrieve, defaults to \code{100}. When you make an API request, you will usually not receive all of the results of that request in a single response. This is because some responses could contain thousands of objects so most responses are paginated by default. \code{previous} fetches the previous page of response (after the initial query) similarly \code{next} fetches the next page and \code{NULL} does not paginate (only makes one query).}
}
\value{
Returns a \code{data.frame} containing targeting sentences as columns.
}
\description{
Retrieves the targeting description of a specific ad or adset.
}
\details{
Corresponds to this API call; \url{https://developers.facebook.com/docs/marketing-api/targeting-description/v2.6}
}
\examples{
\dontrun{
# get information on account
ads <- grabAds(account.id = "act_123456789012345", token = "XXXXXXXXXXX")

# pick random ad.id
rand_id <- sample(ads$id, 1)

# fetch targeting description of random ad
findTarget(id = rand_id, n = 100, token = "XXXXXXXXXXX")
}

}
\author{
John Coene \email{john.coene@cmcm.com}
}
\seealso{
\code{\link{grabAds}}
}

