% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkTargetSentence.R
\name{checkTargetSentence}
\alias{checkTargetSentence}
\title{Check object target}
\usage{
checkTargetSentence(id, token, n = 100, limit = 100)
}
\arguments{
\item{id}{ID of object to retrieve.}

\item{token}{A valid token as returned by \code{\link{fbAuthenticate}} or a short-term token from \href{https://developers.facebook.com/tools/explorer}{facebook Graph API Explorer}.}

\item{n}{Number of results to retrieve, defaults to \code{100}. When you make an API request, you will usually not receive all of the results of that request in a single response. This is because some responses could contain thousands of objects so most responses are paginated by default. \code{previous} fetches the previous page of response (after the initial query) similarly \code{next} fetches the next page and \code{NULL} does not paginate (only makes one query).}

\item{limit}{Number of results requested at each API call, defaults to 100.
Sometimes useful to bring it down if many results (\code{n}) are required as the 
API might otherwise return \code{error_code: 1} or in other words an
 "Unknown error".}
}
\value{
Returns a \code{data.frame} containing targeting sentences as columns.
}
\description{
Retrieves the targeting description (in sentence lines) 
of a specific ad or adset. See \code{\link{checkTargetTree}}
}
\details{
Corresponds to this API call; \url{https://developers.facebook.com/docs/marketing-api/targeting-description/v2.7}
}
\examples{
\dontrun{
# get information on account
ads <- grabAds(account.id = "act_123456789012345", token = "XXXXXXXXXXX")

# pick random ad.id
rand_id <- sample(ads$id, 1)

# fetch targeting description of random ad
checkTargetSentence(id = rand_id, n = 100, token = "XXXXXXXXXXX")
}

}
\seealso{
\code{\link{checkTargetTree}}
}
\author{
John Coene \email{jcoenep@gmail.com}
}
