% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkTargetTree.R
\name{checkTargetTree}
\alias{checkTargetTree}
\title{Check object target tree}
\usage{
checkTargetTree(id, token, limit.type = NULL, limit = 100, n = 100)
}
\arguments{
\item{id}{ID of object to retrieve.}

\item{token}{A valid token as returned by \code{\link{fbAuthenticate}} or a short-term token from \href{https://developers.facebook.com/tools/explorer}{facebook Graph API Explorer}.}

\item{limit.type}{Limit the type of audience to retrieve, defaults to \code{NULL}. See \code{\link{findParams}}}

\item{limit}{Number of results requested at each API call, defaults to 100.
Sometimes useful to bring it down if many results (\code{n}) are required as the 
API might otherwise return \code{error_code: 1} or in other words an
 "Unknown error".}

\item{n}{Number of results to retrieve, defaults to \code{100}. When you make an API request, you will usually not receive all of the results of that request in a single response. This is because some responses could contain thousands of objects so most responses are paginated by default. \code{previous} fetches the previous page of response (after the initial query) similarly \code{next} fetches the next page and \code{NULL} does not paginate (only makes one query).}
}
\description{
Retrieves the targeting description of a specific object.
}
\details{
Corresponds to this API call; \url{https://developers.facebook.com/docs/marketing-api/targeting-description/v2.6}
}
\examples{
\dontrun{
# get information on account
ads <- grabAds(account.id = "act_123456789012345", token = "XXXXXXXXXXX")

# fetch targeting description of random ad
checkTargetTree(id = sample(ads$id, 1), token = "XXXXXXXXXXX")
}

}
\author{
John Coene \email{john.coene@cmcm.com}
}
\seealso{
\code{\link{checkTargetSentence}}
}

