% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkUsers.R
\name{checkUsers}
\alias{checkUsers}
\title{Check users}
\usage{
checkUsers(id, token)
}
\arguments{
\item{id}{ID of object to retrieve.}

\item{token}{A valid token as returned by \code{\link{fbAuthenticate}} or a short-term token from \href{https://developers.facebook.com/tools/explorer}{facebook Graph API Explorer}.}
}
\description{
Retrieves the users having access to the account
}
\examples{
\dontrun{
# get information on account
ads <- grabAds(account.id = "act_123456789012345", token = "XXXXXXXXXXX")
}

}
\author{
John Coene \email{john.coene@cmcm.com}
}
\seealso{
\code{\link{checkTargetTree}}
}

