% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createAudience.R
\name{createAudience}
\alias{createAudience}
\title{Create an audience}
\usage{
createAudience(token, account.id, description, name, subtype = "CUSTOM")
}
\arguments{
\item{token}{A valid token as returned by \code{\link{fbAuthenticate}} 
or a short-term token from 
\href{https://developers.facebook.com/tools/explorer}{facebook Graph API Explorer}.}

\item{account.id}{The id of the account you want to retrieve (Required),
see \href{https://www.facebook.com/business/help/1492627900875762}{how to find yours}.}

\item{description}{Description of the audience}

\item{name}{Name of the audience}

\item{subtype}{Type of custom audience, derived from original data source.
See \code{\link{findParams}}.}
}
\description{
Create an empty audience
}
\details{
To create a custom audience you'll first need to create a 
blank audience. Then, you'll want to add people to the blank audience you j
ust created by updating the users edge of the audience using 
\code{\link{fillAudience}}.
}
\author{
John Coene \email{john.coene@cmcm.com}
}
\seealso{
\code{\link{fillAudience}}
}

