% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createAudienceUser.R
\name{createAudienceUser}
\alias{createAudienceUser}
\title{Add users to custom audience}
\usage{
createAudienceUser(token, data, schema = "MADID_SHA256")
}
\description{
Add users to custom audience
}
\author{
John Coene \email{john.coene@cmcm.com}
}

