% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createLabels.R
\name{createLabels}
\alias{createLabels}
\title{Create labels}
\usage{
createLabels(account.id, labels, token)
}
\arguments{
\item{account.id}{ID of the account to retrieve labels from.}

\item{labels}{Name of labels to create}

\item{token}{A valid token as returned by \code{\link{fbAuthenticate}} or a short-term token from \href{https://developers.facebook.com/tools/explorer}{facebook Graph API Explorer}.}
}
\value{
Vector of label IDs created
}
\description{
Create labels in given account
}
\examples{
\dontrun{
# run authentication with your app details
TK <- fbAuthenticate(app.id = "1234567890123456", 
                     app.secret = "16xx79321xx0130x2x10a08x3e2x80xx")
                     
# get 100 accounts users (me) has access to
accounts <- listAccounts("me", TK)

ids <- createLabels(accs$id[1], labels = c("first label", "second label"),
                    token = TK)
}

}
\author{
John Coene \email{john.coene@cmcm.com}
}
\seealso{
\code{\link{getLabels}}
}

