% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillAudience.R
\name{fillAudience}
\alias{fillAudience}
\title{Fill a custom audience}
\usage{
fillAudience()
}
\description{
Add people to custom audience.
}
\author{
John Coene \email{john.coene@cmcm.com}
}
\seealso{
\code{\link{createAudience}}
}

