% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_action_breakdowns.R
\name{find_action_breakdowns}
\alias{find_action_breakdowns}
\title{find_action_breakdowns}
\usage{
find_action_breakdowns()
}
\description{
Returns all valid \code{action_breakdowns} which can be used in the get-family functions. No input, see usage.
}
\author{
John Coene <john.coene@cmcmc.com>
}
\seealso{
\code{\link{get_account}}
}

