% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_breakdowns.R
\name{find_breakdowns}
\alias{find_breakdowns}
\title{find_breakdowns}
\usage{
find_breakdowns()
}
\description{
Returns all valid \code{breakdwowns} which can be used in the get-family functions.. No input, see usage.
}
\author{
John Coene <john.coene@cmcmc.com>
}
\seealso{
\code{\link{get_account}}
}

