% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_date_preset.R
\name{find_date_preset}
\alias{find_date_preset}
\title{find_date_preset}
\usage{
find_date_preset()
}
\description{
Returns all valid values for \code{date_preset} which can be used in the get-family functions. No input, see usage.
}
\author{
John Coene <john.coene@cmcmc.com>
}
\seealso{
\code{\link{get_account}}
}

