% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findActionBreakdowns.R
\name{findActionBreakdowns}
\alias{findActionBreakdowns}
\title{findActionBreakdowns}
\usage{
findActionBreakdowns()
}
\description{
Returns all valid \code{action_breakdowns} which can be used in the get-family functions. No input, see usage.
}
\author{
John Coene <john.coene@cmcmc.com>
}
\seealso{
\code{\link{getAny}}
}

