% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findAds.R
\name{findAds}
\alias{findAds}
\title{findAds}
\usage{
findAds(id, token, fields = "default", n = 100, verbose = FALSE)
}
\arguments{
\item{id}{Your ad account id, starting by "act_" and followed by 15 digits (Required), see \href{https://www.facebook.com/business/help/1492627900875762}{how to find yours}.}

\item{token}{A valid token as returned by \code{\link{fbAuthenticate}} or a short-term token from \href{https://developers.facebook.com/tools/explorer}{facebook Graph API Explorer}.}

\item{fields}{default (\code{"default"}) returns the most popular ones. Run \code{\link{findFields}} to see all valid fields.}

\item{n}{Number of results to retrieve, defaults to \code{100}. When you make an API request, you will usually not receive all of the results of that request in a single response. This is because some responses could contain thousands of objects so most responses are paginated by default. \code{previous} fetches the previous page of response (after the initial query) similarly \code{next} fetches the next page and \code{NULL} does not paginate (only makes one query).}

\item{verbose}{Defaults to \code{FALSE} if \code{TRUE} will print information on the query in the console.}
}
\description{
Fetches the list of ads under a given account, campaign or adset.
}
\examples{
\dontrun{
adsets <- findAdsets(account.id = "act_123456789123456", token = "XXXXXX")

ads <- findAds(account.id = adsets$id[1], token = "XXXXXX")

# use info for query
dat <- getAd(ad.id = sample(ads$id, 1), token = "XXXXXX")
}

}
\author{
John Coene <john.coene@cmcm.com>
}
\seealso{
\code{\link{getAd}}, \code{\link{findAdsets}}
}

