% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findDatePreset.R
\name{findDatePreset}
\alias{findDatePreset}
\title{findDatePreset}
\usage{
findDatePreset()
}
\description{
Returns all valid values for \code{date_preset} which can be used in the get-family functions. No input, see usage.
}
\examples{
\dontrun{
# get information on account
info <- findInfo(account.id = "act_123456789012345", token = "XXXXXXXXXXX")

# take random ad.id
set.seed(123)
rand_id <- sample(info$campaigns$id, 1)

# get date.preset
date <- findDatePreset()[9]

# fetch Campaign data for the last 30 days broken down by days (rather than aggregate)
data <- getAny(id = rand_id, token = "XXXXXXXXXXX", date.preset = date, time.increment = 1)
}

}
\author{
John Coene <john.coene@cmcm.com>
}
\seealso{
\code{\link{getAny}}
}

