% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findFields.R
\name{findFields}
\alias{findFields}
\title{findFields}
\usage{
findFields()
}
\description{
Returns all 73 valid \code{fields} which can be used in the get-family functions. No input, see details for more information.
}
\details{
fields vary from \code{age} to \code{cpm} for the full list please see the online official \href{https://developers.facebook.com/docs/marketing-api/reference/ads-insights/}{documentation}
}
\examples{
\dontrun{
# video-related metrics
videos <- findFields()[grep("video", findFields())]

# get account data on fields of interest
data <- getAccount(account.id = "act_123456789012345", token = "XXXXXXXXXXXX", fields = videos)
}

}
\author{
John Coene <john.coene@cmcmc.com>
}
\seealso{
\code{\link{getAccount}}
}

