% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findInfo.R
\name{findInfo}
\alias{findInfo}
\title{findInfo}
\usage{
findInfo(account.id, token)
}
\arguments{
\item{account.id}{Your ad account id, starting by "act_" and followed by 15 digits (Required), see \href{https://www.facebook.com/business/help/1492627900875762}{how to find yours}.}

\item{token}{A valid token as returned by \code{\link{fbAuthenticate}} or a short-term token from \href{https://developers.facebook.com/tools/explorer}{facebook Graph API Explorer}.}
}
\value{
Returns a list of four data.frames.
\itemize{
\item ads - campaign id, adset id, id, name and bid amount of each ad (rows)
\item adsets - id, name and campaign_id of each adset (rows)
\item campaigns - id and name of all campaigns
\item summary - includes all of the above in one neat \code{data.frame} using ads as rows.
}

See list structure (\code{str}) or \code{names}
}
\description{
Fetches the list of campaigns, adsets and ads under a given account.
}
\details{
Essentially consists of three separate API calls thus may take a few seconds; \code{Sys.sleep} of 3 seconds between each query.
One call is made for each object, ads, adsets and campaigns respectively.
}
\examples{
\dontrun{
lst <- getInfo(account.id = "act_123456789123456", token = "XXXXXX")

# inspect structure of feedback
str(lst)

# show names of data.frames within list
names(lst)

# summary list
summary <- lst$summary
}

}
\author{
John Coene <john.coene@cmcmc.com>
}
\seealso{
\code{\link{getAny}}
}

