% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findParams.R
\name{findParams}
\alias{findParams}
\title{Find valid parameters}
\usage{
findParams(params = c("level", "breakdowns", "date.preset",
  "action.breakdowns", "action.report.time", "action.attribution.windows",
  "time.increment"))
}
\arguments{
\item{params}{parameter(s) to check}
}
\value{
If one \code{params} is passed the function returns a 
\code{vector} otherwise a \code{list} is returned.
}
\description{
Helper function to check valid values for parameters of 
GET-family functions (i.e.: \code{\link{getAny}}).
}
\details{
Function applies to parameters of the following GET-family functions 
(i.e.: \code{\link{getAny}}) parameters:
\itemize{
\item \code{level}
\item \code{breakdowns}
\item \code{date.preset}
\item \code{time.increment}
\item \code{ation.breakdowns}
\item \code{action.report.time}
\item \code{action.attribution.windows}
\item \code{subtype}
\item \code{limit.type}
}
}
\examples{
\dontrun{

# check all valid fields
parameters <- findParams()

# check valid breakdowns
val_br <- findParams("breakdowns")

# use parameter in GET-family function
dat <- getAny(id = "act_123456789012345", token = "XXXXXXXXXXXXX",
              breakdowns = val_br[1])

}

}
\author{
John Coene \email{john.coene@cmcm.com}
}
\seealso{
\code{\link{getAny}}, \code{\link{checkTargetTree}}
}

