% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findStatus.R
\name{findStatus}
\alias{findStatus}
\title{findStatus}
\usage{
findStatus(id, token, fields = "effective_status")
}
\arguments{
\item{id}{The id of the campaign, adset or ad you want to retrieve
(Required), see 
\href{https://www.facebook.com/business/help/1492627900875762}{how to find yours}.}

\item{token}{A valid token as returned by \code{\link{fbAuthenticate}}
or a short-term token from 
\href{https://developers.facebook.com/tools/explorer}{facebook Graph API Explorer}.}

\item{fields}{default (\code{"effective_status"}) 
returns the most popular ones. Run \code{\link{findFields}} 
to see all valid fields.}
}
\description{
Returns effective as well as configured status and created time, name, account id of object requested (campaign, adset or ad).
}
\details{
\itemize{
\item \code{configured_status}: The configured status of the object (campaign, adset or ad). If object is campaign and this status is \code{PAUSED}, all its active ad sets and ads will be paused and have an effective status \code{CAMPAIGN_PAUSED}.
\item \code{effective_status} The effective status of this campaign/adset/ad; \code{enum {ACTIVE, PAUSED, DELETED, PENDING_REVIEW, DISAPPROVED, PREAPPROVED, PENDING_BILLING_INFO, CAMPAIGN_PAUSED, ARCHIVED, ADSET_PAUSED}}
}
}
\examples{
\dontrun{
get information on account
ads <- findAds(id = "act_123456789012345", token = "XXXXXXXXXXX")

# take random ad.id
rand_id <- sample(ads$id, 1)

# getStatus
(status <- findStatus(id = rand_id, token = "XXXXXXXXXX"))
}

}
\author{
John Coene <john.coene@cmcm.com>
}
\seealso{
\code{\link{findAds}}
}

