% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_account.R
\name{get_account}
\alias{get_account}
\title{get_account}
\usage{
get_account(account_id, fields = "default",
  action_attribution_windows = NULL, action_breakdowns = NULL,
  action_report_time, breakdowns, date_preset, level, time_increment,
  time_range, token)
}
\arguments{
\item{account_id}{Your ad account id, starting by "act_" and followed by 15 digits, see \href{https://www.facebook.com/business/help/1492627900875762}{how to find yours}.}

\item{fields}{There are in total 73 valid fields for this function. Run \code{\link{find_fields}} to print see all valid fields.}

\item{action_attribution_windows}{Determines what is the attribution window for the actions. For example, \code{c("28d_click")} means the API returns all actions that happened 28 days after someone clicked on the ad (Optional).}

\item{action_breakdowns}{How to break down action results. Supports more than one breakdowns (Optional).}

\item{action_report_time}{Not Yet Implemented}

\item{breakdowns}{Not Yet Implemented}

\item{date_preset}{Not Yet Implemented}

\item{level}{Not Yet Implemented}

\item{time_increment}{Not Yet Implemented}

\item{time_range}{Not Yet Implemented}

\item{token}{A valid token as returned by \code{\link{fb_authenticate}} or a short-term token from \href{https://developers.facebook.com/tools/explorer}{facebook Graph API Explorer}.}
}
\description{
Get insights on advertising performance on ad account. See \href{https://developers.facebook.com/docs/marketing-api/reference/ad-account/insights/}{documentation} for more information
}
\author{
John Coene <john.coene@cmcm.com>
}

