% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getActivities.R
\name{getActivities}
\alias{getActivities}
\title{getActivities}
\usage{
getActivities(token, account.id, fields = "default", since = NULL,
  until = NULL, n = 100, limit = 100)
}
\arguments{
\item{token}{A valid token as returned by \code{\link{fbAuthenticate}} or a 
short-term token from 
\href{https://developers.facebook.com/tools/explorer}{facebook Graph API Explorer}.}

\item{account.id}{The id of the account you want to retrieve (Required),
see \href{https://www.facebook.com/business/help/1492627900875762}{how to find yours}.}

\item{fields}{There are numerous valid fields defaults to 
(\code{default}) which returns the most popular ones. See 
\code{\link{findFields}}}

\item{since}{date \code{YYYY-MM-DD}}

\item{until}{date \code{YYYY-MM-DD}, defaults to now}

\item{n}{Number of results to retrieve, defaults to \code{100}. 
When you make an API request, you will usually not receive all of the 
results of that request in a single response. 
This is because some responses could contain thousands of objects so 
most responses are paginated by default.}

\item{limit}{Number of results requested at each API call, defaults to \code{100}.
Sometimes useful to bring it down if many results (\code{n}) are required as the 
API might otherwise return \code{error_code: 1} or in other words an
 "Unknown error".}
}
\description{
Get activities on an ad account.
}
\examples{
\dontrun{
# run authentication with your app details
fbOAuth <- fbAuthenticate(app.id = "1234567890123456", 
                          app.secret = "76xx79121xx0130x2x10a08x3e2x80xx", 
                          scope = "ads_management")

# get account ids
acc <- grabAccounts(id = "me", token = "XXXXXXXXXXX")

# get activity
act <- getActivities(token = fbOAuth, account.id = sample(act, 1), n = 200)
}

}
\author{
John Coene \email{john.coene@cmcm.com}
}
\seealso{
\code{\link{fbAuthenticate}}, \code{\link{grabAccounts}}
}

