% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAd.R
\name{getAd}
\alias{getAd}
\title{getAd}
\usage{
getAd(ad.id, fields = "default", action.attribution.windows = NULL,
  action.breakdowns = NULL, action.report.time = NULL, breakdowns = NULL,
  date.preset = NULL, level = NULL, time.increment = NULL,
  time.range = NULL, n = 100, token)
}
\arguments{
\item{ad.id}{The ad id (Required), see \href{https://www.facebook.com/business/help/1492627900875762}{how to find yours}.}

\item{fields}{There are in total 73 valid fields; default (\code{NULL}) returns the most popular ones. Run \code{\link{findFields}} to see all valid fields.}

\item{action.attribution.windows}{Determines what is the attribution window for the actions. For example, \code{c("28d_click")} means the API returns all actions that happened 28 days after someone clicked on the ad (Optional). See details below for valid values.}

\item{action.breakdowns}{How to break down action results. Supports more than one breakdowns (Optional). Run \code{\link{findActionBreakdowns}} to see all valid action breakdowns.}

\item{action.report.time}{Determines the report time of action stats. For example, if a person saw the ad on Jan 1st but converted on Jan 2nd, when you query the API with \code{action.report.time="impression"}, you will see a conversion on Jan 1st. When you query the API with \code{action.report.time="conversion"}, you will see a conversion on Jan 2nd (Optional).}

\item{breakdowns}{How to break down the result. Does not support more than one breakdown, except \code{c("age", "gender")} and \code{"impression_device", "placement"}. The option \code{impression_device} cannot be used by itself (Optional). Run \code{\link{findBreakdowns}} to see all valid breakdowns.}

\item{date.preset}{Represents a relative time range (Optional). This field is ignored if \code{time.range} is specified. Run \code{\link{findDatePreset}} to see all valid presets.}

\item{level}{Represents the level of result (Optional). Must be one  of \code{ad}, \code{adset}, \code{campaign}, \code{account}.}

\item{time.increment}{If it is an integer, it is the number of days from 1 to 90. After you pick a reporting period by using \code{time.range} or \code{date.preset}, you may choose to have the results for the whole period, or have results for smaller time slices. If "all_days" is used, it means one result set for the whole period. If "monthly" is used, you will get one result set for each calendar month in the given period. Or you can have one result set for each N-day period specified by this param.}

\item{time.range}{time range must be \code{c(since = 'YYYY-MM-DD', until='YYYY-MM-DD')}}

\item{n}{Number of results to retrieve, defaults to \code{100}. When you make an API request, you will usually not receive all of the results of that request in a single response. This is because some responses could contain thousands of objects so most responses are paginated by default. \code{previous} fetches the previous page of response (after the initial query) similarly \code{next} fetches the next page and \code{NULL} does not paginate (only makes one query).}

\item{token}{A valid token as returned by \code{\link{fbAuthenticate}} or a short-term token from \href{https://developers.facebook.com/tools/explorer}{facebook Graph API Explorer}.}
}
\description{
Get insights on advertising performances of an ad. See \href{https://developers.facebook.com/docs/marketing-api/reference/adgroup/insights/}{documentation} for more information
}
\details{
This function refers to the following API call \url{https://developers.facebook.com/docs/marketing-api/reference/ad-account/insights/},
it is strongly encouraged to have a look a the latter link.
only the following parameters are not available \code{default_summary}, \code{filtering}, 
\code{summary}, \code{sort} and \code{time_ranges}.

Valid \code{action.attribution.windows}:
\itemize{
\item 1d_view
\item 7d_view
\item 28d_view
\item 1d_click
\item 7d_click
\item 28d_click
}
}
\author{
John Coene <john.coene@cmcm.com>
}
\seealso{
\code{\link{fbAuthenticate}}
}

