% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCreative.R
\name{getCreative}
\alias{getCreative}
\title{Get creative content}
\usage{
getCreative(object.story.id, token, fields = c("name", "id", "message"))
}
\arguments{
\item{object.story.id}{Object story id, can be obtained with 
\code{\link{listCreatives}}}

\item{token}{A valid token as returned by \code{\link{fbAuthenticate}} or a 
short-term token from 
\href{https://developers.facebook.com/tools/explorer}{facebook Graph API Explorer}.}

\item{fields}{Variables to retrieve, see \code{\link{findFields}}}
}
\description{
Get creative content details and meta-data
}

