% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEstReach.R
\name{getEstReach}
\alias{getEstReach}
\title{Get ad estimated reach}
\usage{
getEstReach(id, token, optimize_for = "APP_INSTALLS")
}
\arguments{
\item{id}{The id of the object you want to retrieve (Required),
see \href{https://www.facebook.com/business/help/1492627900875762}{how to find yours}.}

\item{token}{A valid token as returned by \code{\link{fbAuthenticate}} or a 
short-term token from 
\href{https://developers.facebook.com/tools/explorer}{facebook Graph API Explorer}.}

\item{optimize_for}{What you are optimizing for. This API generates min, 
median, and max bid estimates related to the optimization goal, defaults to 
\code{APP_INSTALLS}, see official documentation or \code{\link{findParams}} 
for valid values}
}
\description{
Retrieves given ad expected reach
}
\examples{
\dontrun{
# run authentication with your app details
TK <- fbAuthenticate(app.id = "1234567890123456", 
                     app.secret = "16xx79321xx0130x2x10a08x3e2x80xx")

# get all accounts users (me) has access to
accounts <- listAccounts("me", TK, n = 999999)
ads <- listAds(sample(accounts$id, 1), TK)

reach <- getEstReach(samples(ads$id, 1), TK)
}

}
\author{
John Coene \email{john.coene@cmcm.com}
}
\seealso{
\code{\link{listLabels}}
}

