% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getImage.R
\name{getImage}
\alias{getImage}
\title{getImage}
\usage{
getImage(id, n = 100, token, verbose = FALSE)
}
\arguments{
\item{id}{The id of the object you want to retrieve (Required), see \href{https://www.facebook.com/business/help/1492627900875762}{how to find yours}. Can be account id, campaign id, adset id or an ad id.}

\item{n}{Number of results to retrieve, defaults to \code{100}. When you make an API request, you will usually not receive all of the results of that request in a single response. This is because some responses could contain thousands of objects so most responses are paginated by default. \code{previous} fetches the previous page of response (after the initial query) similarly \code{next} fetches the next page and \code{NULL} does not paginate (only makes one query).}

\item{token}{A valid token as returned by \code{\link{fbAuthenticate}} or a short-term token from \href{https://developers.facebook.com/tools/explorer}{facebook Graph API Explorer}.}

\item{verbose}{Defaults to \code{FALSE} if \code{TRUE} will print information on the query in the console.}
}
\value{
Returns IDs and URLs of all images contained in object (single row if ad.id is passed as arguement)
}
\description{
Fetches URL(s) name(s) and ID(s) of images used for an ad, an adset, a campaign or all the images in an account.
}
\details{
Call \href{https://developers.facebook.com/docs/marketing-api/reference/adgroup/adcreatives/}{adcreatives}.
}
\examples{
\dontrun{
# authenticate
fbOAuth <- fbAuthenticate(app.id = "1234567890123456", 
                          app.secret = "16xx79321xx0130x2x10a08x3e2x80xx", 
                          scope = "ads_management")
# get account ids
act <- findAccounts(id = "me", token = fbOAuth)

# get all images in ad account
img_acc <- getImage(id = act[2,2], token = fbOAuth)

# get campaigns, adsets and ads IDs from account
info <- findObjects(account.id = act[2,2], token = token = fbOAuth)

# get all images in random campaign
img_camp <- getImage(id = sample(info$summary$campaign_id, 1), token = fbOAuth)

# get images used in an adset
img_adset <- getImage(sid = sample(info$adsset$id, 1), token = fbOAuth)

# get ad image
img_ad <- getImage(id = sample(info$ads$id, 1), token = fbOAuth)
}

}
\author{
John Coene <john.coene@cmcm.com>
}
\seealso{
\code{\link{fbAuthenticate}}, \code{\link{findAccounts}}, \code{\link{findObjects}}
}

