% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLabCampaigns.R
\name{getLabCampaigns}
\alias{getLabCampaigns}
\title{Get campaigns by labels}
\usage{
getLabCampaigns(account.id, labels, fields = c("id", "name"),
  operator = "ANY", token, n = 100, verbose = FALSE)
}
\arguments{
\item{account.id}{ID of the account to retrieve labels from.}

\item{labels}{IDs of labels to retrieve.}

\item{fields}{Fields (variables) to retrieve, see \code{\link{findFields}} 
for valid values, defaults to \code{id} and \code{name}.}

\item{operator}{Matching operator either \code{ALL} or \code{ANY}, defaults to \code{ANY}}

\item{token}{A valid token as returned by \code{\link{fbAuthenticate}} or a short-term token from \href{https://developers.facebook.com/tools/explorer}{facebook Graph API Explorer}.}

\item{n}{Number of results to return}

\item{verbose}{Defaults to \code{FALSE} if \code{TRUE} will print information on the 
queries in the console.}
}
\description{
Retrieves campaigns by labels under given account.
}
\examples{
\dontrun{
# run authentication with your app details
TK <- fbAuthenticate(app.id = "1234567890123456", 
                     app.secret = "16xx79321xx0130x2x10a08x3e2x80xx")

# get all accounts users (me) has access to
accounts <- listAccounts("me", TK, n = 999999)

set.seed(19880525) # set seed for sample() reproducability

# get labels of random account
labs <- listLabels(account.id = sample(accounts$id, 1), token = TK)

# get campaigns by labels
camps <- getLabCampaigns(sample(accounts$id, 1), labels = labs$id, 
                         token = TK)
}

}
\seealso{
\code{\link{listLabels}}
}
\author{
John Coene \email{jcoenep@gmail.com}
}
