% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLabels.R
\name{getLabels}
\alias{getLabels}
\title{Get labels}
\usage{
getLabels(account.id, token)
}
\arguments{
\item{account.id}{ID of the account to retrieve labels from.}

\item{token}{A valid token as returned by \code{\link{fbAuthenticate}} or a short-term token from \href{https://developers.facebook.com/tools/explorer}{facebook Graph API Explorer}.}
}
\description{
Retrieves ad labels under given account.
}
\examples{
\dontrun{
# get all accounts
accounts <- listAccounts("me", token = "XXXXXXXXXXX")

# get labels of random account
labs <- getLabels(account.id = sample(accounts$id, 1), token = "XXXXXXXXXXX")
}

}
\author{
John Coene \email{john.coene@cmcm.com}
}
\seealso{
\code{\link{getLabCampaigns}}
}

