% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grabAds.R
\name{grabAds}
\alias{grabAds}
\title{Grab list of ads}
\usage{
grabAds(id, token, fields = "default", ..., n = 100, verbose = FALSE,
  limit = 100)
}
\arguments{
\item{id}{ID of object to grab from. 
\href{https://www.facebook.com/business/help/1492627900875762}{how to find yours}.}

\item{token}{A valid token as returned by \code{\link{fbAuthenticate}}
 or a short-term token from 
\href{https://developers.facebook.com/tools/explorer}{facebook Graph API Explorer}.}

\item{fields}{default (\code{"default"}) returns the most popular ones.
Run \code{\link{findFields}} to see all valid fields.}

\item{...}{additional parameters to pass to insights.}

\item{n}{Number of results to retrieve, defaults to \code{100}.
When you make an API request, you will usually not receive all
 of the results of that request in a single response. 
 This is because some responses could contain thousands of objects
  so most responses are paginated by default. \code{previous} 
  fetches the previous page of response (after the initial query) 
  similarly \code{next} fetches the next page and \code{NULL} does not 
  paginate (only makes one query).}

\item{verbose}{Defaults to \code{FALSE} if \code{TRUE} 
will print information on the query in the console.}

\item{limit}{Number of results requested at each API call, defaults to 
\code{100}.}
}
\description{
Fetches the list of ads under a given account, campaign or adset.
}
\examples{
\dontrun{
ads <- grabAds(id = "act_123456789012345", token = "XXXXXX")
}

}
\author{
John Coene \email{john.coene@cmcm.com}
}
\seealso{
\code{\link{getAd}}
}

