% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grabCampaigns.R
\name{grabCampaigns}
\alias{grabCampaigns}
\title{grabCampaigns}
\usage{
grabCampaigns(id, token, fields = "default", ..., n = 100,
  verbose = FALSE)
}
\arguments{
\item{id}{Your ad account id, starting by "act_" and followed by 15 digits
(Required), see 
\href{https://www.facebook.com/business/help/1492627900875762}{how to find yours}.}

\item{token}{A valid token as returned by \code{\link{fbAuthenticate}}
 or a short-term token from 
\href{https://developers.facebook.com/tools/explorer}{facebook Graph API Explorer}.}

\item{fields}{default (\code{"default"}) returns the most popular ones.
Run \code{\link{findFields}} to see all valid fields.}

\item{...}{additional parameters to pass to insights.}

\item{n}{Number of results to retrieve, defaults to \code{100}.
When you make an API request, you will usually not receive all
 of the results of that request in a single response. 
 This is because some responses could contain thousands of objects
  so most responses are paginated by default. \code{previous} fetches the previous page of response (after the initial query) similarly \code{next} fetches the next page and \code{NULL} does not paginate (only makes one query).}

\item{verbose}{Defaults to \code{FALSE} if \code{TRUE} 
will print information on the query in the console.}
}
\description{
Fetches the list of campaigns under a given account.
}
\examples{
\dontrun{
camp <- findCampaigns(account.id = "123456789", token = "XXXXXX")

# use info for query
dat <- getCampaign(campaign.id = sample(camp$id, 1), date.preset = "last_week",
 token = "XXXXX")
}

}
\author{
John Coene <john.coene@cmcm.com>
}
\seealso{
\code{\link{getCampaign}}
}

