% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listApps.R
\name{listApps}
\alias{listApps}
\title{List Applications}
\usage{
listApps(id, token, n = 100, verbose = FALSE, limit = 100)
}
\arguments{
\item{id}{The id of the object you want to grab from (Required), see 
\href{https://www.facebook.com/business/help/1492627900875762}{how to find yours}.}

\item{token}{A valid token as returned by \code{\link{fbAuthenticate}} 
or a short-term token from 
\href{https://developers.facebook.com/tools/explorer}{facebook Graph API Explorer}.}

\item{n}{Number of results to retrieve, defaults to \code{100}. 
When you make an API request, you will usually not receive all of the 
results of that request in a single response. This is because some 
responses could contain thousands of objects so most responses are 
paginated by default.}

\item{verbose}{Defaults to \code{FALSE} if \code{TRUE} will print information on the 
queries in the console.}

\item{limit}{Number of results requested at each API call, defaults to 
\code{100}.}
}
\description{
Fetches the list of applications under a given account.
}
\examples{
\dontrun{
(apps <- listApps(id = "act_123456789012345", token = "XXXXXX"))
}

}
\author{
John Coene \email{john.coene@cmcm.com}
}

