% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listBusinesses.R
\name{listBusinesses}
\alias{listBusinesses}
\title{listBusinesses}
\usage{
listBusinesses(token, id = "me", n = 100, verbose = FALSE, limit = 100)
}
\arguments{
\item{token}{A valid token as returned by \code{\link{fbAuthenticate}} or a short-term token from \href{https://developers.facebook.com/tools/explorer}{facebook Graph API Explorer}.}

\item{id}{user ID, defaults to \code{me}.}

\item{n}{Number of results to retrieve, defaults to \code{100}. When you make an API request, you will usually not receive all of the results of that request in a single response. This is because some responses could contain thousands of objects so most responses are paginated by default. \code{previous} fetches the previous page of response (after the initial query) similarly \code{next} fetches the next page and \code{NULL} does not paginate (only makes one query).}

\item{verbose}{Defaults to \code{FALSE} if \code{TRUE} will print information on the query in the console.}

\item{limit}{Number of results requested at each API call, defaults to 100.}
}
\description{
Fetches all businesses under the authenticated user. 
Requires access to Business Managers' API.
}
\details{
Returns data.frame of account IDs and status. requires \code{ads_management} permission!
}
\examples{
\dontrun{
# get account ids
accs <- listBusinesses(token = "XXXXXX")
}

}
\author{
John Coene \email{john.coene@cmcm.com}
}

