% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{optIt}
\alias{optIt}
\title{get options}
\usage{
optIt(params, fp, ff, fct = "getAny")
}
\arguments{
\item{params}{parameters to test}

\item{fp}{\code{\link{findParams}}}

\item{ff}{\code{\link{findFields}}}

\item{fct}{Function passed from parent FUN}
}
\description{
get options
}

