% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{scopeCheckC}
\alias{scopeCheckC}
\title{TestScope}
\usage{
scopeCheckC(scope, valid)
}
\arguments{
\item{scope}{Scope to test}

\item{valid}{Valid scopes to test against}
}
\description{
TestScope
}

