% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateAdset.R
\name{updateAdset}
\alias{updateAdset}
\title{Update adset}
\usage{
updateAdset(id, settings, token)
}
\arguments{
\item{id}{ID of object to update.}

\item{settings}{\code{list} of settings to update}

\item{token}{A valid token as returned by \code{\link{fbAuthenticate}} or a short-term token from \href{https://developers.facebook.com/tools/explorer}{facebook Graph API Explorer}.}
}
\value{
Boolean, success \code{TRUE} or \code{FALSE}
}
\description{
Update a specific adset
}
\examples{
\dontrun{
# run authentication with your app details
TK <- fbAuthenticate(app.id = "1234567890123456", 
                     app.secret = "16xx79321xx0130x2x10a08x3e2x80xx")
                     
# get 100 accounts users (me) has access to
accounts <- listAccounts("me", TK)

ids <- createLabels(accs$id[1], labels = "my label", token = TK)

adset_ids <- grabAdsets(accs$id[1], TK) # retrieve adsets

# collapse ids
ids <- paste0("'", paste0(ids, collapse = "', '"), "'")
                  
# create settings to update - ad label                   
my_settings <- list(adlabels = paste0("[{id:", ids,"}]")) 

updateAdset(adset_ids[1], my_settings, TK)
}

}
\author{
John Coene \email{john.coene@cmcm.com}
}

