% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbAuthenticate.R
\name{fbAuthenticate}
\alias{fbAuthenticate}
\title{fbAuthenticate}
\usage{
fbAuthenticate(app.id, app.secret, scope = "ads_read",
  app.name = "optional")
}
\arguments{
\item{app.id}{ID of your facebook app. \href{https://developers.facebook.com/apps}{See facebook apps}}

\item{app.secret}{App secret of facebook app. \href{https://developers.facebook.com/apps}{See facebook apps}}

\item{scope}{General permissions, either "ads_management" or "ads_read", defaults to the latter, \href{https://developers.facebook.com/docs/facebook-login/permissions}{see permissions}.}

\item{app.name}{Name of your app, optional.}
}
\description{
Authenticate through Facebook Graph API with your facebook application which you can create at \href{https://developers.facebook.com/apps}{https://developers.facebook.com/apps}
}
\details{
There are two different ways of making authenticated requests. One is to obtain
a temporary access token from \url{https://developers.facebook.com/tools/explorer}, 
the other is to use a long-term token provided by this very function. 

Either way you will have to create an app; go to \url{https://developers.facebook.com/apps}, register as a developer
and create a new app. You will also need a verified Facebook account.

To get the short-term token:

\enumerate{
\item Go to facebook's Graph API Explorer: \url{https://developers.facebook.com/tools/explorer}
\item Select your app in the top right corner (defaults to "Graph API Explorer")
\item Then in the dropdown underneath hit "Get Token" and select "Get User Access Token"
\item In the popup navigate to "Extended Permissions" and select "ads_management" and/or "ads_read"
\item Click "Get Access Token" and allow the app in the following popup
\item Copy and Save the token in your R sesssion.
}
However, this token has a 2-hour lifetime by default and thus
needs to be renewed frequently. The alternative is to create an OAuth token using this function and your "App ID" and "App Secret"
}
\examples{
 
\dontrun{
# run authentication with your app details
fbOAuth <- fb_authenticate(app.id = "1234567890123456", 
                           app.secret = "76xx79121xx0130x2x10a08d3e2x80xx", 
                           scope = "ads_management")
print(fbOAuth)
<Token>
<oauth_endpoint>
 authorize: https://www.facebook.com/dialog/oauth
 access:    https://graph.facebook.com/oauth/access_token
<oauth_app> optional
 key:    1234567890123456
 secret: <hidden>
 <credentials> access_token, expires
---
}

}
\author{
John coene \email{john.coene@cmcm.com}
}

