% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_fields.R
\name{find_fields}
\alias{find_fields}
\title{find_fields}
\usage{
find_fields()
}
\description{
Returns all 73 valid \code{fields} which can be used in the get-family functions. No input, see usage.
}
\author{
John Coene <john.coene@cmcmc.com>
}
\seealso{
\code{\link{get_account}}
}

