% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findActionBreakdowns.R
\name{findActionBreakdowns}
\alias{findActionBreakdowns}
\title{findActionBreakdowns}
\usage{
findActionBreakdowns()
}
\description{
Returns all valid \code{action_breakdowns} which can be used in the get-family functions. No input, see usage.
}
\examples{
\dontrun{
# get information on account
obj <- findObjects(account.id = "act_123456789012345", token = "XXXXXXXXXXX")

# take random ad.id
set.seed(123)
rand_id <- sample(obj$ads$id, 1)

# carousel-related action_breakdowns
carousel <- findActionBreakdowns()[grep("carousel", findActionBreakdowns())]

# fetch Ad data broken down by actions related to carousel
data <- getAny(id = rand_id, token = "XXXXXXXXXXX", action.breakdowns = carousel)
}

}
\author{
John Coene <john.coene@cmcm.com>
}
\seealso{
\code{\link{getAny}}, \code{\link{findObjects}}
}

