% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findBreakdowns.R
\name{findBreakdowns}
\alias{findBreakdowns}
\title{findBreakdowns}
\usage{
findBreakdowns()
}
\description{
Returns all valid \code{breakdwowns} which can be used in the get-family functions.. No input, see usage.
}
\author{
John Coene <john.coene@cmcmc.com>
}
\seealso{
\code{\link{getAny}}
}

