% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findBreakdowns.R
\name{findBreakdowns}
\alias{findBreakdowns}
\title{findBreakdowns}
\usage{
findBreakdowns()
}
\description{
Returns all valid \code{breakdwowns} which can be used in the get-family functions.. No input, see usage.
}
\examples{
\dontrun{
# get information on account
info <- findInfo(account.id = "act_123456789012345", token = "XXXXXXXXXXX")

# take random ad.id
set.seed(123)
rand_id <- sample(info$adset$id, 1)

# date.preset
date <- findDatePreset()[grep("7",  findDatePreset())]

# fetch Adset data broken down by age and gender for the last 7 days
data <- getAny(id = rand_id, token = "XXXXXXXXXXX", breakdowns = c("age", "gender"), date.preset = date)
}

}
\author{
John Coene <john.coene@cmcmc.com>
}
\seealso{
\code{\link{getAny}}, \code{\link{findInfo}}, \code{\link{findDatePreset}}
}

