% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findDatePreset.R
\name{findDatePreset}
\alias{findDatePreset}
\title{findDatePreset}
\usage{
findDatePreset()
}
\description{
Returns all valid values for \code{date_preset} which can be used in the get-family functions. No input, see usage.
}
\author{
John Coene <john.coene@cmcmc.com>
}
\seealso{
\code{\link{getAny}}
}

