% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findFields.R
\name{findFields}
\alias{findFields}
\title{findFields}
\usage{
findFields(fct)
}
\arguments{
\item{fct}{Name of the function you need the fields for}
}
\description{
Returns all valid \code{fields} which can be used in the get-family functions.
See details for more information.
}
\details{
fields vary from \code{age} to \code{cpm}
Applies to the following functions (\code{fct} argument)
\itemize{
\item \code{\link{getAny}}
\item \code{\link{getAccount}}
\item \code{\link{getAdset}}
\item \code{\link{getAd}}
\item \code{\link{findCampaigns}}
\item \code{\link{findAdsets}}
\item \code{\link{findAds}}
\item \code{\link{findStatus}}
}
}
\examples{
\dontrun{
# video-related metrics
videos <- findFields("getAccount")[grep("video", findFields("getAccount"))]

# get account data on fields of interest
data <- getAccount(account.id = "act_123456789012345", token = "XXXXXXXXXXXX", 
fields = videos)
}

}
\author{
John Coene <john.coene@cmcm.com>
}
\seealso{
\code{\link{getAccount}}
}

