% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findFields.R
\name{findFields}
\alias{findFields}
\title{findFields}
\usage{
findFields(fct = "getAny")
}
\arguments{
\item{fct}{Name of the function you need the fields for}
}
\description{
Returns all valid \code{fields} which can be used in 
the GET-family functions.
See details below for more information.
}
\details{
fields vary from \code{date} to \code{cpm}
Applies to the following functions (\code{fct} argument, i.e.: 
\code{fct = "whatStatus"})
\itemize{
\item \code{\link{getAny}}
\item \code{\link{getAccount}}
\item \code{\link{getAdset}}
\item \code{\link{getAd}}
\item \code{\link{getImage}}
\item \code{\link{getCreative}}
\item \code{\link{getActivity}}
\item \code{\link{grabCreatives}}
\item \code{\link{grabAccounts}}
\item \code{\link{grabCampaigns}}
\item \code{\link{grabAdsets}}
\item \code{\link{grabAds}}
\item \code{\link{checkStatus}}
}
}
\examples{
\dontrun{
# video-related metrics
videos <- findFields("getAccount")[grep("video", findFields("getAccount"))]

# get account data on fields of interest
data <- getAccount(account.id = "act_123456789012345", token = "XXXXXXXXXXXX", 
                   fields = videos)
}

}
\author{
John Coene \email{john.coene@cmcm.com}
}
\seealso{
\code{\link{getAccount}}
}

