% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_account.R
\name{get_account}
\alias{get_account}
\title{get_account}
\usage{
get_account(account_id, fields = "default",
  action_attribution_windows = NULL, action_breakdowns = NULL,
  action_report_time = NULL, breakdowns = NULL, date_preset = NULL,
  level = NULL, time_increment = NULL, time_range = NULL,
  paginate = "next", token)
}
\arguments{
\item{account_id}{Your ad account id, starting by "act_" and followed by 15 digits, see \href{https://www.facebook.com/business/help/1492627900875762}{how to find yours}.}

\item{fields}{There are in total 73 valid fields for this function. Run \code{\link{find_fields}} to print see all valid fields.}

\item{action_attribution_windows}{Determines what is the attribution window for the actions. For example, \code{c("28d_click")} means the API returns all actions that happened 28 days after someone clicked on the ad (Optional).}

\item{action_breakdowns}{How to break down action results. Supports more than one breakdowns (Optional).}

\item{action_report_time}{Determines the report time of action stats. For example, if a person saw the ad on Jan 1st but converted on Jan 2nd, when you query the API with \code{action_report_time="impression"}, you will see a conversion on Jan 1st. When you query the API with \code{action_report_time="conversion"}, you will see a conversion on Jan 2nd.}

\item{breakdowns}{How to break down the result. Does not support more than one breakdown, except \code{c("age", "gender")} and \code{"impression_device", "placement"}. The option \code{impression_device} cannot be used by itself.}

\item{date_preset}{Represents a relative time range. This field is ignored if \code{time_range} or \code{time_ranges} is specified.}

\item{level}{Represents the level of result. Must be one of ad, adset, campaign, account.}

\item{time_increment}{If it is an integer, it is the number of days from 1 to 90. After you pick a reporting period by using \code{time_range} or \code{date_preset}, you may choose to have the results for the whole period, or have results for smaller time slices. If "all_days" is used, it means one result set for the whole period. If "monthly" is used, you will get one result set for each calendar month in the given period. Or you can have one result set for each N-day period specified by this param.}

\item{time_range}{time range must be \code{c(since = 'YYYY-MM-DD', until='YYYY-MM-DD')}}

\item{token}{A valid token as returned by \code{\link{fb_authenticate}} or a short-term token from \href{https://developers.facebook.com/tools/explorer}{facebook Graph API Explorer}.}
}
\description{
Get insights on advertising performance on ad account. See \href{https://developers.facebook.com/docs/marketing-api/reference/ad-account/insights/}{documentation} for more information
}
\details{
This function refers to the following API call \url{https://developers.facebook.com/docs/marketing-api/reference/ad-account/insights/},
it is strongly encouraged to have a look a the latter link.
only the following parameters are not available \code{default_summary}, \code{filtering}, 
\code{summary}, \code{sort} and \code{time_ranges}.
}
\author{
John Coene <john.coene@cmcm.com>
}
\seealso{
\code{\link{fb_authenticate}}
}

