% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCreative.R
\name{getCreative}
\alias{getCreative}
\title{Get creative content}
\usage{
getCreative(object.story.id, token, fields = c("name", "id", "message"))
}
\arguments{
\item{object.story.id}{Object story id, can be obtained with 
\code{\link{listCreatives}}}

\item{token}{A valid token as returned by \code{\link{fbAuthenticate}} or a 
short-term token from 
\href{https://developers.facebook.com/tools/explorer}{facebook Graph API Explorer}.}

\item{fields}{Variables to retrieve, see \code{\link{findFields}}}
}
\description{
Get creative content details and meta-data
}
\examples{
\dontrun{
# run authentication with your app details
fbOAuth <- fbAuthenticate(app.id = "1234567890123456", 
                          app.secret = "76xx79121xx0130x2x10a08x3e2x80xx", 
                          scope = "ads_management")
ls <- listCreatives(id = "act_1231313", token = fbOAuth)
(creative <- getCreative(object.story.id = sample(ls$object_story_id, 1), 
                         token = fbOAuth))
}

}
\author{
John Coene \email{jcoenep@gmail.com}
}

